safari-600.7-branch/LayoutTests/inspector-obsolete/styles/styles-overriden-properties-expected.txt

Tests that overriding shorthands within rule are visible.

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#main  { (styles-overriden-properties.html:4)
/-- overloaded --/ background: #000;
background: #bada55;

[expanded] 
div  { (user agent stylesheet)
display: block;
